services/stream-processor/src/hafnium_stream/processors/enrichment.py
"""
Transaction Enrichment Processor

Enriches transactions with customer profile and velocity features.
"""

from datetime import datetime, timezone
from typing import Any, Dict

from faust import Table


async def enrich_transaction(
    event: Dict[str, Any],
    velocity_table: Table,
) -> Dict[str, Any]:
    """
    Enrich a transaction with additional features.
    
    Adds:
    - Customer profile data
    - Velocity features (transaction counts, sums)
    - Network features (counterparty analysis)
    
    Args:
        event: Raw transaction event
        velocity_table: Faust table with velocity counters
    
    Returns:
        Enriched transaction event
    """
    customer_id = event.get("customer_id")
    
    # Get velocity features from state
    velocity = velocity_table.get(customer_id, {
        "txn_count_24h": 0,
        "txn_sum_24h": 0.0,
    })
    
    # Build enriched event
    enriched = {
        **event,
        "customer_profile": await get_customer_profile(customer_id),
        "velocity_features": {
            "txn_count_24h": velocity.get("txn_count_24h", 0),
            "txn_sum_24h": velocity.get("txn_sum_24h", 0.0),
            "txn_avg_24h": (
                velocity.get("txn_sum_24h", 0.0) / max(velocity.get("txn_count_24h", 1), 1)
            ),
        },
        "network_features": await get_network_features(
            customer_id,
            event.get("counterparty_id"),
        ),
        "enriched_at": datetime.now(timezone.utc).isoformat(),
    }
    
    return enriched


async def get_customer_profile(customer_id: str) -> Dict[str, Any]:
    """
    Fetch customer profile from database or cache.
    
    In production, this would query Redis cache first,
    then fall back to database.
    """
    # Placeholder - integrate with actual profile store
    return {
        "risk_tier": "medium",
        "days_since_onboarding": 180,
        "total_txn_count": 150,
        "avg_txn_amount": 500.0,
        "kyc_status": "verified",
    }


async def get_network_features(
    customer_id: str,
    counterparty_id: str,
) -> Dict[str, Any]:
    """
    Compute network-based features.
    
    Analyzes the relationship between customer and counterparty
    in the transaction graph.
    """
    # Placeholder - integrate with graph database
    return {
        "counterparty_risk_score": 0.2,
        "shared_counterparty_count": 0,
        "is_first_interaction": False,
    }
